Refactored cli usage of archive into a seperate plugin
